Criando pacote 2 SSIS

+++ DatawarehouseSucos/Tabelas/Fato_002.sql
@@ -4,7 +4,7 @@
     [Cod_Produto] NVARCHAR(50) NOT NULL, 
     [Cod_Fabrica] NVARCHAR(50) NOT NULL, 
     [Cod_Dia] NVARCHAR(50) NOT NULL, 
-    [Frete] DECIMAL(18, 4) NULL, 
+    [Frete] FLOAT NULL, 
     PRIMARY KEY ([Cod_Cliente], [Cod_Produto], [Cod_Fabrica], [Cod_Dia]), 
     CONSTRAINT [FK_Fato_002_Dim_Cliente] FOREIGN KEY (Cod_Cliente) REFERENCES Dim_Cliente(Cod_Cliente), 
     CONSTRAINT [FK_Fato_002_Dim_Produto] FOREIGN KEY (Cod_Produto) REFERENCES Dim_Produto(Cod_Produto),
