+++ DatawarehouseSucos/Tabelas/Fato_004.sql
@@ -4,7 +4,7 @@
     [Cod_Produto] NVARCHAR(50) NOT NULL, 
     [Cod_Organizacional] NVARCHAR(50) NOT NULL, 
     [Cod_Dia] NVARCHAR(50) NOT NULL, 
-    [Meta_Faturamento] DECIMAL(18, 4) NULL, 
+    [Meta_Faturamento] FLOAT NULL, 
     PRIMARY KEY ([Cod_Cliente], [Cod_Produto], [Cod_Organizacional], [Cod_Dia]), 
     CONSTRAINT [FK_Fato_004_Dim_Cliente] FOREIGN KEY (Cod_Cliente) REFERENCES Dim_Cliente(Cod_Cliente), 
     CONSTRAINT [FK_Fato_004_Dim_Produto] FOREIGN KEY (Cod_Produto) REFERENCES Dim_Produto(Cod_Produto),
